+++ src/tests/uttilities/providerMockData.ts
@@ -1,5 +1,5 @@
-import { RpcProviderTestCase } from "../types";
-import NetworkConfig from '../config/network-config.json';
+import { RpcProviderTestCase } from "../../types";
+import NetworkConfig from '../../config/network-config.json';
 const ProxyReaderAddress = NetworkConfig.networks[1].contracts.ProxyReader.address.toLowerCase();
 
 export const RpcProviderTestCases: RpcProviderTestCase = [
